Use installed CLI for integration testing
- runs the CLI under Python 2.7 for testing
- CLI has been moved inside the Python package
- explicit dependencies added on required libraries
- new `--identity` option to pass in the SSH key to use
- integration tests automatically install the CLI
