+++ src/leappto.egg-info/entry_points.txt
@@ -1,0 +1,3 @@
+[console_scripts]
+leapp-tool = leappto.cli:main
+
